Align `OAuth2RequestPerformer.perform` return type `URLSession.data`
